Created Edit Quote
Added the ability to edit pending quotes for contractors. Currently it only works for newly created quotes, due to the missing contractorID. I added contractorID to jobQuote, and newly created quotes assign the ID properly. I also added a  nav button to editJob and rerouted the user to the same page after submitting changes to see them after the change.